requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
elevenlabs==1.9.0
pydub==0.25.1
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0
loguru==0.7.3
